fix: update endpoint for partial update

+++ apps-api/backoffice/src/modules/admin/banner/repository.ts
@@ -1,9 +1,11 @@
 import { FilePath } from '@pple-today/api-common/dtos'
 import { FileService, PrismaService } from '@pple-today/api-common/services'
 import { fromRepositoryPromise } from '@pple-today/api-common/utils'
-import { BannerNavigationType, BannerStatusType } from '@pple-today/database/prisma'
+import { BannerNavigationType } from '@pple-today/database/prisma'
 import Elysia from 'elysia'
 import { err, ok } from 'neverthrow'
+
+import { CreateBannerBody, UpdateBannerBody } from './models'
 
 import { FileServicePlugin } from '../../../plugins/file'
 import { PrismaServicePlugin } from '../../../plugins/prisma'
@@ -32,13 +34,7 @@ export class AdminBannerRepository {
     )
   }
 
-  async createBanner(data: {
-    imageFilePath: FilePath
-    navigation: BannerNavigationType
-    destination: string
-    startAt: Date
-    endAt: Date
-  }) {
+  async createBanner(data: CreateBannerBody) {
     const moveFileResult = await fromRepositoryPromise(
       this.fileService.$transaction(async (fileTx) => {
         const moveResult = await fileTx.bulkMoveToPublicFolder([data.imageFilePath])
@@ -59,14 +55,24 @@ export class AdminBannerRepository {
         })
 
         return tx.banner.create({
-          data: {
-            imageFilePath: newFileName,
-            navigation: data.navigation,
-            destination: data.destination,
-            order: lastBanner ? lastBanner.order + 1 : 1,
-            startAt: data.startAt,
-            endAt: data.endAt,
-          },
+          data:
+            data.navigation === BannerNavigationType.MINI_APP
+              ? {
+                  imageFilePath: newFileName,
+                  navigation: data.navigation,
+                  order: lastBanner ? lastBanner.order + 1 : 1,
+                  startAt: data.startAt,
+                  endAt: data.endAt,
+                  miniApp: { connect: { id: data.miniAppId } },
+                }
+              : {
+                  imageFilePath: newFileName,
+                  navigation: data.navigation,
+                  order: lastBanner ? lastBanner.order + 1 : 1,
+                  startAt: data.startAt,
+                  endAt: data.endAt,
+                  destination: data.destination,
+                },
           select: {
             id: true,
           },
@@ -83,19 +89,11 @@ export class AdminBannerRepository {
     return ok(createBannerResult.value)
   }
 
-  async updateBannerById(
-    id: string,
-    data: {
-      imageFilePath: FilePath
-      navigation: BannerNavigationType
-      destination: string
-      status: BannerStatusType
-    }
-  ) {
+  async updateBannerById(id: string, data: UpdateBannerBody) {
     const existingBanner = await fromRepositoryPromise(
       this.prismaService.banner.findUniqueOrThrow({
         where: { id },
-        select: { imageFilePath: true },
+        select: { imageFilePath: true, navigation: true },
       })
     )
 
@@ -103,6 +101,9 @@ export class AdminBannerRepository {
 
     const updateFileResult = await fromRepositoryPromise(
       this.fileService.$transaction(async (fileTx) => {
+        if (data.imageFilePath === undefined) {
+          return existingBanner.value.imageFilePath as FilePath
+        }
         if (existingBanner.value.imageFilePath !== data.imageFilePath) {
           const deleteResult = await fileTx.deleteFile(
             existingBanner.value.imageFilePath as FilePath
@@ -130,6 +131,15 @@ export class AdminBannerRepository {
           navigation: data.navigation,
           destination: data.destination,
           status: data.status,
+          startAt: data.startAt,
+          endAt: data.endAt,
+          miniApp:
+            data.miniAppId &&
+            (data.navigation === BannerNavigationType.MINI_APP ||
+              (data.navigation === undefined &&
+                existingBanner.value.navigation === BannerNavigationType.MINI_APP))
+              ? { connect: { id: data.miniAppId } }
+              : undefined,
         },
       })
     )
@@ -140,7 +150,7 @@ export class AdminBannerRepository {
       return err(updateBannerResult.error)
     }
 
-    return ok(updateBannerResult.value)
+    return ok()
   }
 
   async deleteBannerById(id: string) {
